chore : redux 설치 및 store 선언
