+++ src/App.jsx
@@ -1,12 +1,7 @@
 import "./App.css";
-import SupabaseTest from "./supabase/SupabaseTest";
 
-function App() {
-  return (
-    <>
-      <SupabaseTest />
-    </>
-  );
+async function App() {
+  return <>{/* <Supabase /> */}</>;
 }
 
 export default App;
