+++ src/redux/slices/supabase.slice.js
@@ -1,0 +1,7 @@
+import getPosts from "../../supabase/Supabase";
+
+const initialState = {
+  dataList: getPosts(),
+};
+
+console.log("initialState", initialState);
