+++ src/redux/store.js
@@ -1,0 +1,7 @@
+import { configureStore } from "@reduxjs/toolkit";
+
+const store = configureStore({
+  reducer: {},
+});
+
+export default store;
